Update README.md
This repository contains my 'DATA ENGINEERING CONCEPTS' outline modeled after (but not strictly aligned with) the flow of the Dataquest Data Engineer course. The outline is in PDF form. It is expected to be modified (primarily with additional information amendments) over time and can serve as a knowledge base framework.

The Jupyter Notebook file ('technique_demo.ipynb') is a "practice space" with code representing many but not all of the various techniques that make up the outline. The Markdown cells in the Notebook provide more specific notations on these technique implementations, strategies, and results. This Notebook can be updated over time without concern for order of operations, as the overall objective is limited to practicing the various techniques in a disparate manner.

The 'synthetic_dataset_primary.csv' and 'high_records.txt' files are a synthetic dataset that is used in many of the coding exercises in the practice space. This dataset has been used in various projects (including the algorithmic development of the dataset itself) and can be found in the Repository 'synthetic-metabolic-dataset.'

This approach represents an iterative and experiential learning method in data engineering. It allows for continuous refinement and expansion of skills through hands-on practice and repeated development cycles. The outline in PDF form provides a structured foundation of data engineering principles, while the Jupyter Notebook offers a practical, interactive space to experiment with these concepts. By combining a formal outline with exploratory coding practices, this project exemplifies a constructivist educational strategy—building knowledge incrementally through direct experimentation, reflection, and adaptation. This method is particularly suited for data engineering, where continuous learning, application, and alignment of theory and practice are essential for mastering complex technical skills.